optimize entity path

+++ src/main/kotlin/de/florianmarsch/spring/fussballmanager/RefrenceDataRestController.kt
@@ -57,16 +57,19 @@ class ReferenceDataRestController {
 
 	@PostMapping("/api/goals/{match}")
 	fun  saveGoals(@RequestBody goals:List<Goal>, @PathVariable(value="match")  match:Int) {
-		logger.info("Save ${goals.size} Goals for ${match}.")
+		logger.debug("Save ${goals.size} Goals for ${match}.")
 		val findByMatch = goalRepo.findByMatch(match)?.filterNotNull()
-		logger.info("Already has ${findByMatch.size} Goals for ${match}.")
+		logger.debug("Already has ${findByMatch.size} Goals for ${match}.")
 		findByMatch?.let {
 			val deleted = it.toMutableList().filter {
 				goals.contains(it).not()
 			}
 
-			logger.info("Need to delete ${deleted.size} Goals for ${match}.")
-			goalRepo.deleteAll(deleted)
+			if(deleted.isEmpty().not()){
+				logger.debug("Need to delete ${deleted.size} Goals for ${match}.")
+				goalRepo.deleteAll(deleted)
+			}
+
 		}
 		goalRepo.saveAll(goals)
 	}
